interfaces implemented.
Created but not yet implemented StateHelper to manage repeated code.
Airport is in its own package. I concidered puting it with the interfaces. and allowing them to directly access its attributes.
	then rather than having a StateHelper. having static methods in AirportStateI do repeated code.
		But as of now a StateHelper seems easier.

Low_Risk is the only implemented State.

im not sure what the "SecurityFactors" mentioned in the assignment description is.
	i was wondering if it is the context. which i have named Airport.